Adds the credit for the picture used in README
I just added the picture credits to the owner so that there
works can be appreciated and if anyone like they can search for
them.
Many of them didn't have the contact details so I wasn't able to
provide it.

I also rectify the mistake I did on README.
Previously it was all bold character but now, I highlighted only the
important words with bold and italic.